backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
secure==0.3.0
pydantic==2.5.3
scikit-learn==1.4.0
web3==6.15.0
python-dotenv==1.0.1
numpy==1.26.3
PyJWT==2.8.0
SQLAlchemy==2.0.25
cryptography==42.0.0
